Modified the ILayouter interface to allow to specify to disable animation steps.
The BaseLayouter contains a protected variable _disableAnimation to toggle this and setter/getter methods to be conform with the Interface.

The AnimatedBaseLayouter is implemented to honor the flag and bypass animations if set. Thus, all layouter derived from that should directly work.

Layouters not using the AnimatedBaseLayouter but still doing animation like Hyperbolic2D would need to be modified (if possible and useful) to honor the flag.