changed the way how the specification of the walking direction was implemented. Despite some disagreement with Ivan Bulanov, I decided to introduce a property of Graph to determine what type of spanning tree should be used. The idea is that for one set of graph data, the decision is only taken once and all spanning trees will then be of the specified type by the graph.
There is one bit of uglyness. Since the type is kept in the Graph object, the GTree
has to query the Graph for its value. However, it also has to query the graph
for other things, so it should not be a very ugly thing.

Generally the code should be much cleaner now, and we have state all over the place
(e.g. autoFit, showHistory, directional, etc. etc.).