removed implicit call to draw() in the following places:
- in VisualGraph.setAllVisible()
- in VisualGraph.currentRootVNode (setter)
- in VisualGraph.updateVisibility()

This means, if you relied on an automatic redraw on setting a new
root node or when changing the visibility (e.g. due to changing the
degree of separation or something similar) then you have to
do something explicit now.

I will see that I make sure a double click will still do a redraw().